requirements.txt
cachelib==0.6.0
certifi==2022.9.24
chardet==4.0.0
click==8.0.4
dataclasses==0.8
epicstore-api==0.1.3
Flask==2.0.3
Flask-Session==0.4.0
fuzzywuzzy==0.18.0
idna==2.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
Levenshtein==0.20.8
MarkupSafe==2.0.1
mysql-connector-python==8.0.31
numpy==1.19.5
pandas==1.1.5
path==16.2.0
protobuf==3.19.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-Levenshtein==0.20.8
pytz==2022.6
rapidfuzz==2.11.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.5.4
secure-smtplib==0.1.1
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.12
Werkzeug==2.0.3
zipp==3.6.0
